added breathe module to psychedelic emergency page
