+++ pages/e/psychedelic.module.scss
@@ -1,0 +1,5 @@
+.breatheSection {
+  & > div {
+    margin-top: 1rem;
+  }
+}
